fix(server): update Materia DTOs and repository to match frontend field naming

+++ server/src/main/java/mx/cucii/school/platform/dto/MateriaRequest.java
@@ -1,9 +1,11 @@
 package mx.cucii.school.platform.dto;
 
+import java.math.BigDecimal;
+
 public record MateriaRequest(
         String nombre,
         String clave,
-        Integer creditos,
+        BigDecimal creditos,
         Integer cuatrimestre,
         Integer planEstudioId
 ) {}
