+++ server/src/main/java/mx/cucii/school/platform/dto/MateriaResponse.java
@@ -1,12 +1,13 @@
 package mx.cucii.school.platform.dto;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
 public record MateriaResponse(
         Integer id,
         String nombre,
         String clave,
-        Integer creditos,
+        BigDecimal creditos,
         Integer cuatrimestre,
         Integer planEstudioId,
         boolean isActive,
